// frontend/src/services/generated/models/CreateVideoUploadResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StoragePresignedUpload } from './StoragePresignedUpload';
import {
    StoragePresignedUploadFromJSON,
    StoragePresignedUploadFromJSONTyped,
    StoragePresignedUploadToJSON,
    StoragePresignedUploadToJSONTyped,
} from './StoragePresignedUpload';
import type { VideoResponse } from './VideoResponse';
import {
    VideoResponseFromJSON,
    VideoResponseFromJSONTyped,
    VideoResponseToJSON,
    VideoResponseToJSONTyped,
} from './VideoResponse';

/**
 * 
 * @export
 * @interface CreateVideoUploadResponse
 */
export interface CreateVideoUploadResponse {
    /**
     * 
     * @type {StoragePresignedUpload}
     * @memberof CreateVideoUploadResponse
     */
    upload?: StoragePresignedUpload;
    /**
     * 
     * @type {VideoResponse}
     * @memberof CreateVideoUploadResponse
     */
    video?: VideoResponse;
}

/**
 * Check if a given object implements the CreateVideoUploadResponse interface.
 */
export function instanceOfCreateVideoUploadResponse(value: object): value is CreateVideoUploadResponse {
    return true;
}

export function CreateVideoUploadResponseFromJSON(json: any): CreateVideoUploadResponse {
    return CreateVideoUploadResponseFromJSONTyped(json, false);
}

export function CreateVideoUploadResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateVideoUploadResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'upload': json['upload'] == null ? undefined : StoragePresignedUploadFromJSON(json['upload']),
        'video': json['video'] == null ? undefined : VideoResponseFromJSON(json['video']),
    };
}

export function CreateVideoUploadResponseToJSON(json: any): CreateVideoUploadResponse {
    return CreateVideoUploadResponseToJSONTyped(json, false);
}

export function CreateVideoUploadResponseToJSONTyped(value?: CreateVideoUploadResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'upload': StoragePresignedUploadToJSON(value['upload']),
        'video': VideoResponseToJSON(value['video']),
    };
}