// frontend/src/services/generated/models/StoragePresignedUpload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StoragePresignedUpload
 */
export interface StoragePresignedUpload {
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof StoragePresignedUpload
     */
    headers?: { [key: string]: string; };
    /**
     * 
     * @type {string}
     * @memberof StoragePresignedUpload
     */
    method?: string;
    /**
     * 
     * @type {string}
     * @memberof StoragePresignedUpload
     */
    url?: string;
}

/**
 * Check if a given object implements the StoragePresignedUpload interface.
 */
export function instanceOfStoragePresignedUpload(value: object): value is StoragePresignedUpload {
    return true;
}

export function StoragePresignedUploadFromJSON(json: any): StoragePresignedUpload {
    return StoragePresignedUploadFromJSONTyped(json, false);
}

export function StoragePresignedUploadFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoragePresignedUpload {
    if (json == null) {
        return json;
    }
    return {
        
        'headers': json['headers'] == null ? undefined : json['headers'],
        'method': json['method'] == null ? undefined : json['method'],
        'url': json['url'] == null ? undefined : json['url'],
    };
}

export function StoragePresignedUploadToJSON(json: any): StoragePresignedUpload {
    return StoragePresignedUploadToJSONTyped(json, false);
}

export function StoragePresignedUploadToJSONTyped(value?: StoragePresignedUpload | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'headers': value['headers'],
        'method': value['method'],
        'url': value['url'],
    };
}